requirements.txt
apipkg==1.4
attrs==17.4.0
Django==2.0.2
execnet==1.5.0
factory-boy==2.10.0
Faker==0.8.11
inflection==0.3.1
pluggy==0.6.0
py==1.5.2
pytest==3.4.0
pytest-django==3.1.2
pytest-factoryboy==2.0.1
pytest-forked==0.2
pytest-xdist==1.22.0
python-dateutil==2.6.1
pytz==2018.3
six==1.11.0
text-unidecode==1.1
